Updates nLinkReacher, pointRobot, mobileBase according to modifications in planarCommon. Unifies velEnv and accEnv to integrate unified verification of observation space.
