+++ pointRobot/envs/acc.py
@@ -1,18 +1,6 @@
-import numpy as np
-from gym import spaces
-
 from pointRobot.envs.pointRobotEnv import PointRobotEnv
+from planarCommon.accEnv import AccEnv
 
 
-class PointRobotAccEnv(PointRobotEnv):
-    def setSpaces(self):
-        o = np.concatenate((self._limUpPos, self._limUpVel))
-        self.observation_space = spaces.Box(low=-o, high=o, dtype=np.float64)
-        self.action_space = spaces.Box(
-            low=-self._limUpAcc, high=self._limUpAcc, dtype=np.float64
-        )
-
-    def continuous_dynamics(self, x, t):
-        vel = x[self._n : self._n * 2]
-        xdot = np.concatenate((vel, self.action))
-        return xdot
+class PointRobotAccEnv(AccEnv, PointRobotEnv):
+    pass
